Eliminate the need for string appending by reversing them
The previous solution would build the complete string for every partial
word which needs to be checked. This is a O(n) operation, executed for
every step of the recursion.
The current solution reverses all strings when the WordSet is built
initially, so this adds an O(n) operation for every word, but only once
at initialization. Later, the word only needs to be reversed when it is
actually a match. Effectively this reduces the amount of O(n) operations
per recursion from n recursions to just as many as there are matches.